--- fashion_bags_commerce/src/component/GlobalStyles/layouts/DefaultLayout/SideBar/Avartar/tesst.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-<body>
-    
-    <img src="../../../../../../../public/logo.png" />
-</body>
-</html>
