+++ fashion_bags_commerce/src/pages/Home/Footer/index.module.scss
@@ -1,5 +1,6 @@
 .footer {
   // position: fixed;
+  max-width: 100%;
   background-color: white;
   .mainFooter {
     .container {
