+++ fashion_bags_commerce/src/pages/Home/InfoCustomer/InformationCustomer/InformationCustomer.scss
@@ -1,0 +1,63 @@
+.infor-custom {
+  width: 1000px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  .infor-custom-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: start;
+    width: 100%;
+    height: 40px;
+    font-size: 16px;
+    .infor-custom-lable-name {
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      width: 200px;
+      font-weight: 600;
+    }
+    .infor-custom-input {
+      width: 100%;
+      .infor-custom-input-item {
+        height: 40px;
+        width: 100%;
+        border-radius: 6px;
+      }
+    }
+  }
+}
+
+.btn-container {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  .btn-back {
+    color: #ffffff;
+    padding: 5px 20px;
+    border-radius: 100px;
+    background-color: gray;
+    cursor: pointer;
+
+    &:hover {
+      background-color: lightgray;
+    }
+  }
+  .btn-update {
+    color: #ffffff;
+    padding: 5px 20px;
+    border-radius: 100px;
+    background-color: #ff5733;
+    cursor: pointer;
+    &:hover {
+      background-color: rgb(248, 107, 107);
+    }
+  }
+}
+.thongTinKhachHang {
+  border: 1px solid #ccc;
+  box-shadow: 5px 5px 5px 5px rgba(0.5, 0.1, 0.1, 0.1);
+  padding: 20px;
+  border-radius: 8px;
+  width: 1200px;
+}
