+++ fashion_bags_commerce/src/pages/Home/InfoCustomer/ProfileView/profile.css
@@ -1,37 +1,44 @@
-/* Trong file CSS hoặc file SCSS của bạn */
 
 .tabLeft{
-    margin-left: 7%;
-    margin-top: 10px;
-    font-size: 30px;
-}
+  margin-left: 7%;
+  margin-top: 15px;
+  font-size: 30px;
+  }
+.infoLabel{
+    background-color: gray;
+    padding: 10px 0 10px 60px;
+    width: 295px;
+  /* position: absolute; */
+
+  }
 .tabLeftChild{
-    margin-left: 90px;
+  margin-left: 90px;
 
 }
 .tabLeftChildRow{
-    font-size: 16px;
+  font-size: 16px;
 
 }
 .tabLeftChildRow:hover{
-    color: blue;
+  color: blue;
 }
 .tabTop {
-  display: flex;
+display: flex;
 
 }
 
 .tabTopChild {
-  flex: 1;
-  text-align: center;
-  font-size: 16px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  margin: 0 5px;
-  padding: 8px;
+flex: 1;
+text-align: center;
+font-size: 16px;
+border: 1px solid #ccc;
+border-radius: 5px;
+margin: 0 5px;
+padding: 8px;
+
 }
 
 .tabTopChild:hover {
-  background-color: #e6e6e6;
-  cursor: pointer;
+background-color: #e6e6e6;
+cursor: pointer;
 }
