+++ fashion_bags_commerce/src/pages/Home/index.scss
@@ -3,8 +3,8 @@ Link {
 }
 .fullpage {
   display: block;
-  width: 100%;
-  float: left;
+  max-width: 100%;
+  // float: left;
   background-color: white;
   text-align: center;
 }
@@ -23,9 +23,8 @@ body .fullpage {
   max-width: 100%;
   margin: 0 auto;
   background-color: white;
-
 }
 
-.header001{
+.header001 {
   background-color: white;
 }
